fix(https_server): fix undeclared `context` error
Compiling the `https_server.cpp` file fails when
`WITH_CPPHTTPLIB_LIB` is not defined.

This is because the `https_free_context` function and `context` vars
are not declared, which results in the following error in GCC 11.3.0:

> ```
> /brski/src/brski/http/https_server.cpp: In function ‘int https_start(http_config*, std::vector<RouteTuple>&, void*, void**)’:
> /brski/src/brski/http/https_server.cpp:33:23: error: ‘context’ was not declared in this scope
>    33 |   https_free_context(*context);
>       |                       ^~~~~~~
> /brski/src/brski/http/https_server.cpp:33:3: error: ‘https_free_context’ was not declared in this scope
>    33 |   https_free_context(*context);
> ```

These functions/param were partially removed in commit
d484024 (feat: Added registrar/server context to routes, 2023-02-01),
but only for the `#ifdef WITH_CPPHTTPLIB_LIB` branch.

Fixes: d4840243ec223ce3ef999cc730e45ca9750bab08