refactor(brski): change funcs to accept `const`
Change some of the functions to accept a `const char *`, instead of just
a plain `char *`.

Unfortunately, since `basename()` can modify the input string in very
rare occasions (see [basename docs][1]), we have to make a copy of the
string before calling `basename()`.

(we're stuck on C++11, so we can't use C++17's
[`std::filesystem::path::filename()`][2])

[1]: https://pubs.opengroup.org/onlinepubs/9699919799/utilities/basename.html
[2]: https://en.cppreference.com/w/cpp/filesystem/path/filename