refactor(brski): declare log_lock_fun as C
Currently, we declare/define the `log_lock_fun` inside C++ code.
We then pass a function pointer to the function to some C code.

This is probably fine on GCC, but on other platforms, this may cause
issues. To be on the safe side, we can declare the function in an
`extern "C" {}` environment.