feat(crypto): add `not_after_absolute` field
Add `not_after_absolute` field to `struct crypto_cert_meta`.

Unlike the existing `not_after` field, which represents an offset from
the current time, the `not_after_absolute` field represents an
ASN.1 GENERALIZEDTIME absolute time.

It can be set to `"99991231235959Z"` for a
[long-lived pledge certificate][rfc8995#2.6.2].

[rfc8995#2.6.2]: https://www.rfc-editor.org/rfc/rfc8995.html#name-infinite-lifetime-of-idevid

BREAKING CHANGE: This is an ABI breaking change for the voucher lib.