fix(crypto): use better OpenSSL error messages
OpenSSL errors are currently just logged using their error code,
using [ERR_get_error()][1]. However, this error code is a very large
number that is non-trivial to decode.

We can use [ERR_error_string()][2] to decode this error code into a
string. I've used `NULL` as the input buffer to use an OpenSSL static
buffer, instead of having to deal with allocating our own buffer.
Technially, this function isn't thread-safe, but considering that we
only call this function on error, this should be okay.

[1]: https://www.openssl.org/docs/man3.1/man3/ERR_get_error.html
[2]: https://www.openssl.org/docs/man3.1/man3/ERR_error_string.html